backend/requirements.txt
# Flask Backend Dependencies
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Database
pymongo>=4.6.0

# AI/ML
openai>=1.0.0
requests>=2.28.0
transformers>=4.30.0

# Data Validation
pydantic>=2.0.0

# Async/WebSocket
websockets>=10.0

# Legacy (can be removed if not used)
groq
